requirements.txt
sentence-transformers
langchain-community
langchain 
langchain-google-genai 
python-dotenv 
twilio